requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
gevent==23.9.1
psutil==5.9.6
llama-cpp-python==0.2.90
typing-extensions
numpy
diskcache